refactor(cli): memory and bench commands move to src/cli/
