+++ bin/lcm.ts
@@ -2,11 +2,13 @@
 import { realpathSync } from "node:fs";
 import { argv, exit, stdin, stdout } from "node:process";
 import { homedir } from "node:os";
-import { join, resolve } from "node:path";
+import { join } from "node:path";
 import { fileURLToPath } from "node:url";
 import { Command, Option } from "commander";
 import { DaemonClient } from "../src/daemon/client.js";
 import { lcmHome, lcmPath } from "../src/lcm-home.js";
+import { registerMemoryCommands } from "../src/cli/memory.js";
+import { registerBenchCommands } from "../src/cli/bench.js";
 
 function readStdin(): Promise<string> {
   return new Promise((resolve) => {
@@ -54,165 +56,6 @@ export function shouldRunMain(invokedPath: string | undefined, currentFilePath: 
 
 
 
-export function registerMemoryCommands(program: Command): void {
-  program
-    .command("search <query>")
-    .description("Search memory across episodic and promoted layers")
-    .option("--limit <n>", "Max results per layer", "5")
-    .option("--layer <name>", "Layer to search: episodic or promoted (repeatable)", collectRepeatedOption, [])
-    .option("--tag <tag>", "Require a tag on matching entries (repeatable)", collectRepeatedOption, [])
-    .helpOption(false)
-    .option("-h, --help", "Show help")
-    .action(async (query: string, opts) => {
-      if (opts.help) {
-        const { printHelp } = await import("../src/cli-help.js");
-        printHelp("search"); exit(0);
-      }
-
-      const layers = normalizeStringList(opts.layer);
-      const tags = normalizeStringList(opts.tag) ?? [];
-      ensureAllowedValues(layers, ["episodic", "promoted"], "--layer");
-
-      const client = await createDaemonClientOrExit();
-      const result = await client.post("/search", {
-        cwd: process.cwd(),
-        query,
-        limit: parsePositiveInteger(String(opts.limit ?? "5"), "--limit"),
-        layers,
-        tags,
-      });
-      printJson(result);
-    });
-
-  program
-    .command("grep <query>")
-    .description("Search raw messages and summaries by keyword or regex")
-    .option("--mode <mode>", "Search mode: full_text or regex", "full_text")
-    .option("--scope <scope>", "Scope: messages, summaries, or both", "both")
-    .option("--since <iso>", "Only include matches on or after this ISO timestamp")
-    .helpOption(false)
-    .option("-h, --help", "Show help")
-    .action(async (query: string, opts) => {
-      if (opts.help) {
-        const { printHelp } = await import("../src/cli-help.js");
-        printHelp("grep"); exit(0);
-      }
-
-      const mode = ensureAllowedValue(opts.mode, ["full_text", "regex"], "--mode");
-      const scope = ensureAllowedValue(opts.scope, ["messages", "summaries", "both"], "--scope");
-
-      const client = await createDaemonClientOrExit();
-      const result = await client.post("/grep", {
-        cwd: process.cwd(),
-        query,
-        mode,
-        scope,
-        since: typeof opts.since === "string" && opts.since.length > 0 ? opts.since : undefined,
-      });
-      printJson(result);
-    });
-
-  program
-    .command("describe <nodeId>")
-    .description("Inspect metadata for a summary or stored memory node")
-    .helpOption(false)
-    .option("-h, --help", "Show help")
-    .action(async (nodeId: string, opts) => {
-      if (opts.help) {
-        const { printHelp } = await import("../src/cli-help.js");
-        printHelp("describe"); exit(0);
-      }
-
-      const client = await createDaemonClientOrExit();
-      const result = await client.post("/describe", { cwd: process.cwd(), nodeId });
-      printJson(result);
-    });
-
-  program
-    .command("expand <nodeId>")
-    .description("Expand a summary node back into source detail")
-    .option("--depth <n>", "Traversal depth", "1")
-    .helpOption(false)
-    .option("-h, --help", "Show help")
-    .action(async (nodeId: string, opts) => {
-      if (opts.help) {
-        const { printHelp } = await import("../src/cli-help.js");
-        printHelp("expand"); exit(0);
-      }
-
-      const client = await createDaemonClientOrExit();
-      const result = await client.post("/expand", {
-        cwd: process.cwd(),
-        nodeId,
-        depth: parsePositiveInteger(String(opts.depth ?? "1"), "--depth"),
-      });
-      printJson(result);
-    });
-
-  program
-    .command("store <text>")
-    .description("Store a durable memory entry for the current project")
-    .option("--tag <tag>", "Attach a tag to the stored memory (repeatable)", collectRepeatedOption, [])
-    .helpOption(false)
-    .option("-h, --help", "Show help")
-    .action(async (text: string, opts) => {
-      if (opts.help) {
-        const { printHelp } = await import("../src/cli-help.js");
-        printHelp("store"); exit(0);
-      }
-
-      const client = await createDaemonClientOrExit();
-      const result = await client.post("/store", {
-        cwd: process.cwd(),
-        text,
-        tags: normalizeStringList(opts.tag) ?? [],
-        metadata: {},
-      });
-      printJson(result);
-    });
-}
-
-function collectRepeatedOption(value: string, previous: string[] = []): string[] {
-  return [...previous, value];
-}
-
-function normalizeStringList(value: unknown): string[] | undefined {
-  if (!Array.isArray(value)) return undefined;
-  const normalized = value.filter((item): item is string => typeof item === "string" && item.length > 0);
-  return normalized.length > 0 ? normalized : undefined;
-}
-
-function parsePositiveInteger(value: string, optionName: string): number {
-  const parsed = Number(value);
-  if (!Number.isFinite(parsed) || parsed <= 0 || !Number.isInteger(parsed)) {
-    console.error(`Invalid ${optionName}: ${value}`);
-    exit(1);
-  }
-  return parsed;
-}
-
-function ensureAllowedValues(values: string[] | undefined, allowed: readonly string[], optionName: string): void {
-  if (!values) return;
-  const invalid = values.filter((value) => !allowed.includes(value));
-  if (invalid.length > 0) {
-    console.error(`Invalid ${optionName}: ${invalid.join(", ")}`);
-    exit(1);
-  }
-}
-
-function ensureAllowedValue(value: unknown, allowed: readonly string[], optionName: string): string | undefined {
-  if (value === undefined) return undefined;
-  if (typeof value !== "string" || !allowed.includes(value)) {
-    console.error(`Invalid ${optionName}: ${String(value)}`);
-    exit(1);
-  }
-  return value;
-}
-
-function printJson(value: unknown): void {
-  stdout.write(JSON.stringify(value, null, 2) + "\n");
-}
-
 let cliDaemonActivity: (() => void) | undefined;
 
 async function admitCliDatabaseWork(): Promise<void> {
@@ -257,61 +100,6 @@ async function createDaemonClientOrExit(spawnTimeoutMs = 5000): Promise<DaemonCl
   }
 
   return new DaemonClient(`http://127.0.0.1:${port}`, tokenPath);
-}
-
-export function registerBenchCommands(program: Command): void {
-  // ─── bench ─────────────────────────────────────────────────────────────────
-  const benchCmd = new Command("bench").description(
-    "Build and run a natural-language retrieval benchmark from this project's ingested sessions",
-  );
-  benchCmd.action(() => { benchCmd.outputHelp(); });
-
-  benchCmd
-    .command("build")
-    .description("Sample ingested sessions and write a local benchmark file")
-    .option("--project <path>", "Project directory (default: cwd)")
-    .option("--n <count>", "Number of questions to generate", "20")
-    .option("--out <file>", "Benchmark file path (default: project memory directory)")
-    .option("--generator <mode>", "Question generator: llm or mechanical", "mechanical")
-    .option("--seed <n>", "Deterministic sampling seed", "42")
-    .option("--language <tag>", "Language to write LLM questions in (BCP 47, e.g. pt-BR); default: detected from the corpus")
-    .action(async (opts) => {
-      const cwd = typeof opts.project === "string" ? resolve(opts.project) : process.cwd();
-      const n = parsePositiveInteger(String(opts.n ?? "20"), "--n");
-      const seed = parsePositiveInteger(String(opts.seed ?? "42"), "--seed");
-      if (!["llm", "mechanical"].includes(opts.generator)) throw new Error("--generator must be llm or mechanical");
-      await admitCliDatabaseWork();
-      const { buildBench } = await import("../src/bench.js");
-      const result = await buildBench({ cwd, n, seed, out: opts.out, generator: opts.generator, language: opts.language });
-      stdout.write(result.stdout);
-      exit(result.exitCode);
-    });
-
-  benchCmd
-    .command("run")
-    .description("Run the benchmark against search and a grep baseline")
-    .option("--project <path>", "Project directory (default: cwd)")
-    .option("--k <n>", "Hit-rate cutoff (default: 5)", "5")
-    .option("--bench-file <file>", "Benchmark file path (default: project memory directory)")
-    .option("--json", "Output structured JSON")
-    .option("--union", "Score against every checkout of this repository, not this project alone")
-    .action(async (opts) => {
-      const cwd = typeof opts.project === "string" ? resolve(opts.project) : process.cwd();
-      const k = parsePositiveInteger(String(opts.k ?? "5"), "--k");
-      await admitCliDatabaseWork();
-      const { runBench } = await import("../src/bench.js");
-      const result = await runBench({
-        cwd,
-        k,
-        benchFile: opts.benchFile,
-        json: opts.json ?? false,
-        union: opts.union ?? false,
-      });
-      stdout.write(result.stdout);
-      exit(result.exitCode);
-    });
-
-  program.addCommand(benchCmd);
 }
 
 async function main() {
@@ -928,7 +716,7 @@ async function main() {
       exit(failures.length > 0 ? 1 : 0);
     });
 
-  registerMemoryCommands(program);
+  registerMemoryCommands(program, { createDaemonClientOrExit });
 
   // ─── diagnose ──────────────────────────────────────────────────────────────
   program
@@ -1246,7 +1034,7 @@ async function main() {
       }
     });
 
-  registerBenchCommands(program);
+  registerBenchCommands(program, { admitCliDatabaseWork });
 
   // ─── promote ───────────────────────────────────────────────────────────────
   program
